requirements.txt
docx-mailmerge==0.3.0
lxml==3.8.0
xlrd==1.1.0
